Split the draco ECS demo into components/entities/systems files
A single scene holding components, entities, and systems does not scale.
Move each component, entity, and system into its own file under
app/ecs/, leaving ecs_scene.rb as the only Conjuration-aware file
(world setup, tick from #update, camera.draw render path).

Cross-directory files use app-root requires (require "app/ecs/...")
because DragonRuby's require_relative cannot traverse "..".
