+++ demo/mygame/app/ecs/components/position.rb
@@ -1,0 +1,4 @@
+class Position < Draco::Component
+  attribute :x, default: 0
+  attribute :y, default: 0
+end
