+++ demo/mygame/app/ecs/components/velocity.rb
@@ -1,0 +1,4 @@
+class Velocity < Draco::Component
+  attribute :dx, default: 0
+  attribute :dy, default: 0
+end
