+++ demo/mygame/app/ecs/entities/critter.rb
@@ -1,0 +1,9 @@
+require "app/ecs/components/position.rb"
+require "app/ecs/components/velocity.rb"
+require "app/ecs/components/sprite.rb"
+
+class Critter < Draco::Entity
+  component Position
+  component Velocity
+  component Sprite
+end
