+++ demo/mygame/app/scenes/ecs_scene.rb
@@ -1,75 +1,6 @@
-class Position < Draco::Component
-  attribute :x, default: 0
-  attribute :y, default: 0
-end
-
-class Velocity < Draco::Component
-  attribute :dx, default: 0
-  attribute :dy, default: 0
-end
-
-class Sprite < Draco::Component
-  attribute :w, default: 24
-  attribute :h, default: 24
-  attribute :r, default: 255
-  attribute :g, default: 255
-  attribute :b, default: 255
-
-  # attr_reader, not a draco attribute: a `{}` default is shared across all entities.
-  attr_reader :primitive
-
-  def after_initialize
-    @primitive = { w: w, h: h, path: :pixel, r: r, g: g, b: b }
-  end
-end
-
-class Critter < Draco::Entity
-  component Position
-  component Velocity
-  component Sprite
-end
-
-class MovementSystem < Draco::System
-  filter Position, Velocity
-
-  def tick(_context)
-    entities.each do |entity|
-      entity.position.x += entity.velocity.dx
-      entity.position.y += entity.velocity.dy
-    end
-  end
-end
-
-class BounceSystem < Draco::System
-  filter Position, Velocity, Sprite
-
-  def tick(context)
-    max_x = context.virtual_w
-    max_y = context.virtual_h
-
-    entities.each do |entity|
-      position = entity.position
-      velocity = entity.velocity
-      sprite = entity.sprite
-
-      if position.x < 0
-        position.x = 0
-        velocity.dx = velocity.dx.abs
-      elsif position.x + sprite.w > max_x
-        position.x = max_x - sprite.w
-        velocity.dx = -velocity.dx.abs
-      end
-
-      if position.y < 0
-        position.y = 0
-        velocity.dy = velocity.dy.abs
-      elsif position.y + sprite.h > max_y
-        position.y = max_y - sprite.h
-        velocity.dy = -velocity.dy.abs
-      end
-    end
-  end
-end
+require "app/ecs/entities/critter.rb"
+require "app/ecs/systems/movement_system.rb"
+require "app/ecs/systems/bounce_system.rb"
 
 class CritterWorld < Draco::World
   systems MovementSystem, BounceSystem
